feat: port the major parsing logic for yaml analyzers to the updated analysis engine
