+++ analysis/testrunner.go
@@ -3,6 +3,7 @@ package analysis
 import (
 	"fmt"
 	"io/fs"
+	"os"
 	"path/filepath"
 	"regexp"
 	"sort"
@@ -144,6 +145,48 @@ func getExpectedIssuesInDir(testDir string, fileFilter func(string) bool) (map[s
 	return expectedIssues, nil
 }
 
+func discoverYamlAnalyzers(testDir string) ([]*Analyzer, error) {
+	var yamlAnalyzers []*Analyzer
+
+	err := filepath.Walk(testDir, func(path string, info fs.FileInfo, err error) error {
+		if err != nil {
+			return nil
+		}
+
+		if info.IsDir() {
+			return nil
+		}
+
+		fileExt := filepath.Ext(path)
+		isYamlFile := fileExt == ".yaml" || fileExt == ".yml"
+		if !isYamlFile {
+			return nil
+		}
+
+		// Check if there's a corresponding test file
+		baseName := strings.TrimSuffix(path, fileExt)
+
+		// Try to read the YAML checker
+		analyzer, err := ReadFromFile(path)
+		if err != nil {
+			// Skip files that aren't valid checkers
+			return nil
+		}
+
+		// Check if corresponding test file exists
+		testFile := baseName + ".test" + GetExtFromLanguage(analyzer.Language)
+		if _, err := os.Stat(testFile); os.IsNotExist(err) {
+			// Skip if no test file exists
+			return nil
+		}
+
+		yamlAnalyzers = append(yamlAnalyzers, &analyzer)
+		return nil
+	})
+
+	return yamlAnalyzers, err
+}
+
 func getExpectedIssuesInFile(file *ParseResult, query *sitter.Query) map[int][]string {
 	commentIdentifier := GetEscapedCommentIdentifierFromPath(file.FilePath)
 
@@ -210,6 +253,13 @@ func RunAnalyzerTests(testDir string, analyzers []*Analyzer) (string, string, bo
 
 	// if there's a test file in the testDir for which there's no analyzer,
 	// it's most likely a YAML checker test, so skip it
+
+	yamlAnalyzers, err := discoverYamlAnalyzers(testDir)
+	if err != nil {
+		return "", "", false, err
+	}
+	analyzers = append(analyzers, yamlAnalyzers...)
+
 	likelyTestFiles := []string{}
 	for _, analyzer := range analyzers {
 		likelyTestFiles = append(likelyTestFiles, fmt.Sprintf("%s.test%s", analyzer.Name, GetExtFromLanguage(analyzer.Language)))
