+++ cmd/globstar/main.go
@@ -16,7 +16,7 @@ func main() {
 
 	cli := cli.Cli{
 		RootDirectory: cwd,
-		Checkers:      nil, // no custom checker set
+		// Checkers:      nil, // no custom checker set
 	}
 
 	err = cli.Run()
